+++ Manatee.Trello/TrelloProcessor.cs
@@ -1,4 +1,7 @@
-﻿using Manatee.Trello.Internal.RequestProcessing;
+﻿using System.Linq;
+using Manatee.Trello.Contracts;
+using Manatee.Trello.Internal.RequestProcessing;
+using Manatee.Trello.Json;
 
 namespace Manatee.Trello
 {
@@ -31,5 +34,22 @@ namespace Manatee.Trello
 		{
 			RestRequestProcessor.CancelPendingRequests();
 		}
+
+		/// <summary>
+		/// Processes webhook notification content.
+		/// </summary>
+		/// <param name="content">The string content of the notification.</param>
+		/// <param name="auth">The <see cref="TrelloAuthorization"/> under which the notification should be processed</param>
+		public static void ProcessNotification(string content, TrelloAuthorization auth = null)
+		{
+			var notification = TrelloConfiguration.Deserializer.Deserialize<IJsonWebhookNotification>(content);
+			var action = new Action(notification.Action, auth);
+
+			foreach (var obj in TrelloConfiguration.Cache.OfType<ICanWebhook>())
+			{
+				obj.ApplyAction(action);
+			}
+		}
+
 	}
 }
